Use small page URL names

+++ index.php
@@ -9,11 +9,11 @@ $MyBBI = new MyBBIntegrator ( $mybb, $db, $cache, $plugins, $lang, $config );
 
 $pagetitle = "SoA - ";
 $pagename = "";
-$pageurl = "";
 if (! isset ( $_REQUEST ['page'] )) {
 	$_REQUEST ['page'] = "index";
 }
 $cleanpageid = str_replace ( ".php", "", preg_replace ( '/\s+/', '', strtolower ( $_REQUEST ['page'] ) ) );
+$pageurl = $cleanpageid;
 switch ($cleanpageid) {
 	case "index" :
 		$pagetitle = "Seed of Andromeda";
@@ -23,22 +23,18 @@ switch ($cleanpageid) {
 	case "thegame" :
 		$pagetitle .= "The Game";
 		$pagename = "The Game.php";
-		$pageurl = "TheGame";
 		break;
 	case "theteam" :
 		$pagetitle .= "The Team";
 		$pagename = "The Team.php";
-		$pageurl = "TheTeam";
 		break;
 	case "screenshots" :
 		$pagetitle .= "Image Media";
 		$pagename = "Screenshots.php";
-		$pageurl = "Screenshots";
 		break;
 	case "videos" :
 		$pagetitle .= "Video Media";
 		$pagename = "Videos.php";
-		$pageurl = "Videos";
 		break;
 	// case "mods" :
 	// $pagetitle = "Mods";
@@ -47,7 +43,6 @@ switch ($cleanpageid) {
 	case "irc" :
 		$pagetitle .= "IRC";
 		$pagename = "IRC.php";
-		$pageurl = "IRC";
 		break;
 	// case "store" :
 	// $pagetitle = "Store";
@@ -57,12 +52,10 @@ switch ($cleanpageid) {
 	case "devlog" :
 		$pagetitle .= "Blogs";
 		$pagename = "Blogs.php";
-		$pageurl = "Blogs";
 		break;
 	case "downloads" :
 		$pagetitle .= "Downloads";
 		$pagename = "Downloads.php";
-		$pageurl = "Downloads";
 		break;
 	// case "reddit" :
 	// $pagetitle .= "Reddit";
@@ -71,24 +64,26 @@ switch ($cleanpageid) {
 	case "blog" :
 		$pagetitle .= "Blog";
 		$pagename = "Blog.php";
-		$pageurl = "Blog";
 		break;
-	case "wip" :
+	case "underconstruction" :
 		$pagetitle = "Under Construction";
 		$pagename = "Under Construction.php";
-		$pageurl = "UnderConstruction";
 		break;
 }
 
-if ($pageurl != "" && $_REQUEST ['page'] != $pageurl) {
+$page_exists = $pagename == "" ? false : file_exists ( "pages/" . $pagename );
+
+if ($pageurl != "" && $_REQUEST ['page'] != $pageurl && $page_exists) {
 	header ( "Location: /" . $pageurl );
 	echo '<html><body><a href="/' . $pageurl . '/>Page moved</a></body></html>';
 	exit ();
 }
-
+if(!$page_exists){
+	$pagetitle .= "Not Found";
+}
 include ("header.php");
 
-if (file_exists ( "pages/" . $pagename )) {
+if ($page_exists) {
 	include ("pages/" . $pagename);
 } else {
 	header ( "HTTP/1.0 404 Not Found" );
